requirements/test.txt
Django==2.2
jsonschema==3.2.0
PyYAML==5.3.1
